🎨 Improve Structure by splitting classes into multiple
