+++ src/Routing/PendingRoutingState.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace ZubZet\Framework\Routing;
+
+class PendingRoutingState {
+
+    public array $middleware = [];
+
+    public function middleware(array $middleware): self {
+        $this->middleware[] = $middleware;
+        return $this;
+    }
+}
+?>
